Changed property to double

+++ Uppgift_Git/JossansKlass.cs
@@ -6,6 +6,6 @@ namespace Uppgift_Git
 {
     class JossansKlass
     {
-        public int Number { get; set; }
+        public double Number { get; set; }
     }
 }
